requirements.txt
algoliasearch-django>=3.0,<4.0
django>=5.0.1
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers